LayoutTests/platform/chromium/http/tests/security/cross-frame-access-location-put-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-location-put-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-location-put.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-location-put-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-location-put.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-location-put-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-location-put.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-location-put-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-location-put.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-location-put-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-location-put.html. Domains, protocols and ports must match.




--------
Frame: '<!--framePath //<!--frame0-->-->'
--------

----- tests for putting window.history and its properties -----

PASS: window.location.assign should be 'function () { [native code] }' and is.
PASS: window.location.replace should be 'function () { [native code] }' and is.
PASS: window.location.reload should be 'function () { [native code] }' and is.
PASS: window.location.toString should be 'function toString() { [native code] }' and is.
PASS: window.location.customAttribute should be 'customAttribute' and is.
